feature: Support audio component Both video component and audio use react-player component. So I created ReactPlayerComponent in lib.
