+++ pynecone/components/media/audio.py
@@ -1,0 +1,8 @@
+"""A audio component."""
+from pynecone.components.libs.react_player import ReactPlayerComponent
+
+
+class Audio(ReactPlayerComponent):
+    """Audio component share with Video component."""
+
+    pass
